std::size_t for sizes + casts

+++ examples/load_stan.hpp
@@ -116,10 +116,10 @@ class DynamicStanModel {
   }
 
   std::size_t unconstrained_dimensions() const {
-    return param_unc_num_(model_ptr_.get());
+    return static_cast<std::size_t>(param_unc_num_(model_ptr_.get()));
   }
   std::size_t constrained_dimensions() const {
-    return param_num_(model_ptr_.get(), true, true);
+    return static_cast<std::size_t>(param_num_(model_ptr_.get(), true, true));
   }
 
   template <typename M>
